pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cfd-demo"
version = "0.1.0"
description = "Modular Lagrangian ocean trajectory simulator (demo)"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Evan Lauer" }]
dependencies = [
    "numpy",
    "matplotlib",
    "netCDF4",
    "typer>=0.12",
    "PyYAML",
    "joblib",
    # optional: "plotly", "streamlit"
]

[project.scripts]
cfd-run = "scripts.run:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["components", "scripts"]
